# 🐎 perform groupby in a sparse way for less memory usage/performance (~250 times faster)