--- a/tests/internal/groupby_test.py
+++ b/tests/internal/groupby_test.py
@@ -1,5 +1,6 @@
 import pyarrow as pa
 import vaex.groupby
+import numpy as np
 
 
 
@@ -14,6 +15,23 @@ def test_combined_grouper():
     assert set(df[grouper_y.binby_expression].tolist()) == {0, 1, 2}
     grouper = vaex.groupby._combine(df, groupers, sort=True)
     assert set(df[grouper.binby_expression].tolist()) == {0, 1, 2}
+
+
+def test_combined_grouper_over64bit():
+    bits = [15, 16, 17] * 2
+    assert sum(bits) > 64
+    N = 2**max(bits)
+    def unique_ints(offset, bit):
+        # create 2**bits unique ints
+        ar = np.full(N, offset, dtype='int32')
+        n = 2**bit
+        ar[:n] = np.arange(offset, offset + n)
+        return ar
+    arrays = {f'x_{i}': unique_ints(i, bit) for i, bit in enumerate(bits)}
+    df = vaex.from_dict(arrays)
+    groupers = [vaex.groupby.Grouper(df[name]) for name in arrays]
+    grouper = vaex.groupby._combine(df, groupers, sort=True)
+    assert df[grouper.binby_expression].nunique() == N
 
 
 def test_combined_grouper_cast():