```python
import numpy as np
import vaex
n = 10_000
x = np.random.rand(n)
y = np.random.rand(n)
z = np.random.rand(n)
df = vaex.from_arrays(x=x, y=y, z=z)
```

before:
```python
%%time
dfg = df.groupby(['x', 'y'], agg=vaex.agg.sum('z'))
#CPU times: user 4.16 s, sys: 5.94 s, total: 10.1 s
#Wall time: 10.1 s
```

after:
```python
%%time
dfg = df.groupby(['x', 'y'], agg=vaex.agg.sum('z'))
# CPU times: user 44.7 ms, sys: 2.52 ms, total: 47.2 ms
# Wall time: 38.7 ms
```

which is about ~250 times faster. A lot of time was wasted on allocating memory as well.

There are situations where this is slower, so we may want to release this with `assume_sparse=False` first.

cc @nicolaskruchten