@maartenbreddels 2021-06-04
Fixes #602
Fixes #929